requirements.txt
SQLAlchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
aiohttp==3.10.10
PyYAML==6.0.2